Update main.c
Fixed code, that has the left and right paddle within the boundaries, and implements a speed change. The speed change doesn't seem to be working as sometimes it feels slower, but after the first hit it definitely does get faster. The score is good as it ends at 10, but it does start at Player 1 as 0, and Player 2 as 1, offering an unfair advantage. Another problem is the ball keeps shooting in the same direction at the start, which is simply not randomized.